tx-base.inc: explicitly set TXBASE variable
Currently the final value of the TXBASE variable is determined by the
'mx6' or 'mx6ul' overrides. In order to change the value of TXBASE it
is necessary to set either TXBASE_mx6 or TXBASE_mx6ul in local.conf
depending on the MACHINE setting.

In order to be able to set TXBASE regardless of the MACHINE type, set
a TXBASE_DEFAULT variable with MACHINE dependent overrides and set
TXBASE from TXBASE_DEFAULT, so TXBASE can easily be overridden in
local.conf.